requirements.txt
Flask==3.0.0
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.1